refactor(MutationOptions): Move `MutationOptions` to its own file
+ Move `MutationOptions` to its own file
+ Extract property checks and conversions from JS to Scala as an extension method of `MutationOptions`
